switch layout to use css grid with breakpoints
