+++ web-ui/src/components/common/Widget.tsx
@@ -1,13 +1,16 @@
+import { classNames } from 'utils'
+
 type Props = {
   title: string
   contents: JSX.Element
+  span?: number
 }
 
-export function Widget({ title, contents }: Props) {
+export function Widget({ title, contents, span }: Props) {
   return (
-    <div>
+    <div className={classNames(span ? `col-span-${span}` : '')}>
       <div className="mb-2 text-center text-xl text-black">{title}</div>
-      <div className="min-h-40 w-full min-w-48 rounded-lg bg-black p-4 text-white shadow-lg">
+      <div className="rounded-lg bg-black p-4 text-white shadow-lg">
         {contents}
       </div>
     </div>
